signed/unsigned warnings on win32

+++ indra/llcommon/llerror.cpp
@@ -620,7 +620,7 @@ namespace LLError
 #endif
 		mFunctionString += std::string(mFunction);
 
-		for (int i = 0; i < tag_count; i++)
+		for (size_t i = 0; i < tag_count; i++)
 		{
             if (strchr(tags[i], ' '))
             {
